# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/DizainerRegister": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "DizainerRegister",
                "description": "DizainerRegister",
                "operationId": "DizainerRegister",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "NULL"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "user"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "format": "surname",
                                        "example": "useryan"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "phone",
                                        "example": "93457898"
                                    },
                                    "phone_code": {
                                        "type": "string",
                                        "format": "phone",
                                        "example": "+374"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "111111"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "111111"
                                    },
                                    "diplom_photo": {
                                        "type": "file",
                                        "format": "file",
                                        "example": "photo.png"
                                    },
                                    "selfi_photo": {
                                        "type": "file",
                                        "format": "file",
                                        "example": "photo.png"
                                    },
                                    "i_agree": {
                                        "type": "string",
                                        "format": "I_agree",
                                        "example": "true"
                                    },
                                    "role_id": {
                                        "type": "string",
                                        "format": "role_id",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/sendCallUser": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "sendCallUser",
                "description": "sendCallUser",
                "operationId": "sendCallUser",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "NULL": {
                                        "type": "string",
                                        "format": "NULL",
                                        "example": "NULL"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/updateRegisterNumber": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "updateRegisterNumber",
                "description": "updateRegisterNumber",
                "operationId": "updateRegisterNumber",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "PhoneCode, phone"
                                ],
                                "properties": {
                                    "PhoneCode": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "+374"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "93789878"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "api/updateveryficode": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "updateveryficode",
                "description": "updateveryficode",
                "operationId": "updateveryficode",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone_veryfi_code": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "4587"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}